requirements_tikz.txt
matplotlib<=3.7
numpy<2
webcolors
tikzplotlib-patched
seaborn
pandas